+++ src/yajlpp/yajlpp.cc
@@ -1584,3 +1584,17 @@ yajlpp_gen::to_string_fragment()
 
     return string_fragment::from_bytes(buf, len);
 }
+
+namespace yajlpp {
+
+auto_mem<yajl_handle_t>
+alloc_handle(const yajl_callbacks* cb, void* cu)
+{
+    auto_mem<yajl_handle_t> retval(yajl_free);
+
+    retval = yajl_alloc(cb, nullptr, cu);
+
+    return retval;
+}
+
+}  // namespace yajlpp